tooling/pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brrtrouter-tooling"
version = "0.1.0"
description = "BRRTRouter-specific development tooling: dependabot automerge, CI helpers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0.3",
    "certifi>=2026.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    # Pinned so local venv, pre-commit (tooling/.venv/bin/ruff), and CI stay aligned.
    "ruff==0.15.9",
    "pre-commit>=4.5.1",
]
# MCP SDK: bounded range so CI resolves a published wheel (not a non-existent floor version).
mcp = [
    "mcp[cli]>=1.27.0,<3",
]

[project.scripts]
brrtrouter = "brrtrouter_tooling.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [".venv", "__pycache__", "build", "dist", "*.egg-info", ".git"]

[tool.ruff.lint]
# World-class: pycodestyle, Pyflakes, bugbear, comprehensions, pyupgrade, simplify,
# isort, pathlib, Ruff, try-except-pass; plus builtins, blind-except, eradicate,
# debugger, errmsg, return, logging, pie, pytest-style, raise, pygrep-hooks,
# pep8-naming, perflint, mccabe, tryceratops
select = [
    "E",     # pycodestyle errors
    "F",     # Pyflakes (undefined, unused, redef)
    "W",     # pycodestyle warnings
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "SIM",   # flake8-simplify
    "I",     # isort
    "PTH",   # flake8-use-pathlib
    "RUF",   # Ruff (ambiguous unicode, unused noqa, etc.)
    "S110",  # try-except-pass
    "A",     # flake8-builtins (shadowing)
    "BLE",   # flake8-blind-except
    "ERA",   # eradicate (commented-out code)
    "T10",   # flake8-debugger
    "EM",    # flake8-errmsg (exception messages)
    "RET",   # flake8-return
    "LOG",   # flake8-logging
    "PIE",   # flake8-pie
    "PT",    # flake8-pytest-style
    "RSE",   # flake8-raise
    "PGH",   # pygrep-hooks (blanket noqa, etc.)
    "N",     # pep8-naming
    "PERF",  # perflint
    "C90",   # mccabe (complexity)
    "TRY",   # tryceratops (exceptions)
]
ignore = [
    "E501",    # line too long
    "S603",    # subprocess: false positives when argv is controlled
    "S607",    # partial executable path: docker/cargo/kubectl are intentional
    "S103",    # chmod 0o755: required for copied binaries
    "T201",    # print: CLI and scripts use print intentionally
    "T203",    # pprint: same
    "PERF203", # try-except in loop: false positives in our patterns
    "PERF402", # manual list copy: prefer clarity in some cases
    "TRY003",  # long exception messages: subjective
    "TRY300",  # try-consider-else: style
    "TRY301",  # abstract raise to inner: style
    "PT011",   # pytest raises too broad: allow for now
    "PT018",   # pytest composite assertion: allow combined asserts in tests for brevity
]

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.lint.per-file-ignores]
# main: arg-parsing dispatcher; complexity expected
"src/brrtrouter_tooling/cli/main.py" = ["C901"]
# Hauliage client CLI (sibling repo hauliage/tooling)
"src/brrtrouter_tooling/workspace/cli/main.py" = ["C901"]
# docker_cmd: argv dispatcher; complexity expected
"src/brrtrouter_tooling/cli/docker_cmd.py" = ["C901"]

# NOTE: Empty print statements to stderr/stdout are forbidden.
# Pattern to avoid: print("", file=sys.stderr) or print("", file=sys.stdout)
# Always provide descriptive error messages. Example:
#   print("Error: GITHUB_REPOSITORY environment variable is required", file=sys.stderr)
# This pattern is checked via code review and should be caught by grep/IDE searches.

[tool.ruff.format]
quote-style = "double"
